extracts server code to internal package
